Create catalog from a queue
We now have two ways to create a catalog. One is from requesting a new catalog from HTTP-Methods and the second is by creating one from imported products. To unify those two ways, I created a queue which triggers a new function ultimately creating a catalog. In OO-Programming you would model this a a call to a separate function